requirements.txt
pyautogen
openai
python-dotenv
pandas
matplotlib
seaborn